+++ app/src/js/components/Requests/index.js
@@ -9,6 +9,7 @@ import InactiveRequestsOverview from './withdrawn';
 import ApprovalStep from './approval';
 import ActionRequestsOverview from './status';
 import EsdisApprovalStep from './esdis-approval';
+import AdditionalReviewQuestion from './additional_review';
 
 const Requests = ({
   location,
@@ -36,6 +37,7 @@ const Requests = ({
               <Route path='/requests/approval/esdis' component={EsdisApprovalStep} />
               <Route path='/requests/approval' component={ApprovalStep} />
               <Route path='/requests/status' component={ActionRequestsOverview} />
+              <Route path='/requests/question' component={AdditionalReviewQuestion} />
             </Switch>
           </div>
         </div>