+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated the My Metrics Page to be a subpage of the Metrics Tab
 - Remove DAAC long name from request details page
 - Adding page for ESDIS final approval
+- Moved additional review question from DAAC Assignment prompt to seperate page
 <!-- Unreleased changes can be added here. -->
 
 ## 1.0.19